Deploy Vault and Salt master on Complector
Summary:
Complector is a dedicated host for Vault and Salt master.

It's not reachable directly from InterNet, you need to use
a bastion, a development server for example, to access it.

It's currently a VM on dreadnought, so it uses the Intranought network.

Ref T1702.

Test Plan: Apply core role on Complector

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1702

Differential Revision: https://devcentral.nasqueron.org/D2615