Ensure /home/username has correct permissions on ZFS servers
Summary:
If the user exists, but the home directory doesn't exist,
a tank/usr/home/luser ZFS volume is created, mounted at
/home/luser, but with root:root ownership.

This changes adds a step to set correctly home directory
permissions. This step will be ignored on non ZFS servers.

As a side-effect, on development servers, if the change runs
again, the permissions will be reset to 0700. This is coherent
with the policy project experiments should be created in /srv.

Test Plan:
  - Delete an user and home directory
  - `state.sls roles/core/users`

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2623