Update Docker socket ownership
Summary:
In D2668, the group name allowed to connect to Docker socket is customized.
Yet, on Dwellers, the `docker` client still can't connect on it.

There is a similar known issue reported upstream, which suggests the socket
systemd unit used for socket-activated services also needs to be aware
of the group.

References:
  - http://0pointer.de/blog/projects/socket-activation.html
  - https://github.com/moby/moby/issues/32611

Ref T1724

Test Plan: `docker ps` on Dwellers

Reviewers: dereckson

Reviewed By: dereckson

Tags: #user-dorianwinty

Maniphest Tasks: T1724

Differential Revision: https://devcentral.nasqueron.org/D2670