Deploy run-php-script wrapper on devserver
Summary:
The `run-php-script` wrapper command searches at several locations
for an executable and run the first one found.

That method allows to prioritize a version local to the project,
installed by Composer, and fallback to the system one if needed.

This concerns tools like phan, psalm, rector.

As phpcs/phpcbf are handled by 69e38320, it's out of scope of the
current wrapper change.

Test Plan: `rector`

Reviewers: DorianWinty, dereckson

Reviewed By: DorianWinty, dereckson

Differential Revision: https://devcentral.nasqueron.org/D2676