WIP: Configure RabbitMQ credentials from Vault
Summary:
RabbitMQ clusters need the following important credentials:
  - Erlang cookie, to allow cluster components to communicate
  - An administrator account to allow maintenance tasks

Both are stored in Vault, cookie can be set in host data volume,
administrator account is set when container is initialized.
A file-as-flag is then created to indicate that's part is done.

References:
  - https://www.rabbitmq.com/clustering.html
  - https://www.rabbitmq.com/rabbitmqctl.8.html#add_user

Ref T752

Test Plan: Provision white-rabbit on docker-002

Reviewers: DorianWinty, dereckson

Reviewed By: dereckson

Maniphest Tasks: T752

Differential Revision: https://devcentral.nasqueron.org/D2792