Deploy tmux-reattach
Summary:
It's in a lot of places, like in /root/bin for docker-001, but not everywhere.
Basically it's a small wrapper to reattach to tmux, and launch it when needed.

tmux 1.8 introduced the `new-session` command with a way to attach a session
if already existing, and we currently have 3.x deployed everywhere. That would
require a deterministic naming approach, to avoid the second fork command call.

So instead, we keep the old pre-1.8 solution to try to attach, and create
the new session when that fails.

Test Plan: `tmux-reattach`

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2850