Allow eggdrop to run Git commands on repositories
Summary:
To gather metadata about deployed versions, `git rev-parse`
can be used to get the hash of the current commit.

This requires ownership on repository or an allowlist.
As repositories don't belong to the eggdrop user, a Git configuration
file with a safe repositories allowlist is generated.

Repositories covered:
  - rVIPER
  - rFBSQL
  - rRABBITMQTCL

Ref T1790

Test Plan: Tested with Wearg

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1790

Differential Revision: https://devcentral.nasqueron.org/D2899