Deploy Relay for SDK development on Dwellers
Summary:
Sentry documentation suggests to use a local Relay installation. Relay source
code (UTSL) indicates the capture mode can be used for that goal.

A `flavour` option is added to the pillar: if set at `dev`, it will install
such Relay deployment and skip geoip volume.

Payloads can then be sent to https://dwellers.nasqueron.org:26300
and results checks in `docker logs sentry_relay` on that server.

We had luck with credentials generation for the production environment,
but we stumbled upon the issue it created an empty file, so it's best
to use a temporary file inside the container, so no need to `docker cp`
or play a pipe dance.

Reference: https://develop.sentry.dev/sdk/basics/

Ref T1790

Test Plan:
Deployed both on docker-002, where it's correctly no-op, and on Dwellers.

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1790

Differential Revision: https://devcentral.nasqueron.org/D2900