Provision Notifications center integration environment
Summary:
This environment provides:

  - Notifications center
  - A separate broker, so we can test new routing and mirror
   traffic without disrupt our existing notifications workflows
  - Vault, as there is plan to use it to store credentials

As the main goal is to integrate and test new features,
"integration" is the chosen name for this environment.

The Vault is for future work for the Notifications center
to retrieve credentials. Credentials used by Salt to deploy
the integration environment uses the regular ops Vault.

Signed-off-by: Sébastien Santoro <dereckson@espace-win.org>

Ref T1755

Test Plan: Deploy to Dwellers

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1755

Differential Revision: https://devcentral.nasqueron.org/D2761