Allow to use _default.conf for PaaS Docker nginx vhost
Summary:
**Current behavior**
A _default.conf standard vhost is provided, but not directly useful
as it needs to be included to a specific <service>.conf file.

**New behavior**
This change detects if <service>.conf exists, and if not, fallback
to _default.conf, so services not needing a specific nginx configuration
can directly use the default one without any need to create a nginx file.

Test Plan:
  - Deploy hound, check it's no-op
  - Deploy Penpot (D2597)

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2968