Allow to restore chmod on shared deployment Git directories
Summary:
Git repositories are normally configured in group-writable,
to be shared by members of the 'ops' groups.

As some Git commands can create a file not-group-writable,
this script allows to ensure permissions. It doesn't touch
to user or world permissions.

Known issue is a file intended to be in read only would get
a strange 464 permission, but no occurence in Git repositories
are known.

Test Plan: Tested on Complector

Reviewers: DorianWinty, dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D2973