Deploy Alkane on webserver-alkane role
Summary:
Deploy and configure the Alkane utility. Alkane is built as FreeBSD
port and package to serve through the Nasqueron packages repository.

It runs under the user `deploy` so it has rights to modify the sites files.

Ref T1826.
Ref T1828.

Test Plan: Deploy on Windriver and test deployment to dummy sites

Reviewers: dereckson

Maniphest Tasks: T1828, T1826

Differential Revision: https://devcentral.nasqueron.org/D3013