Create "web" and web domains groups
Summary:
The "web" group is moved to webserver-core, as a part of the effort
to sync all our roles nginx configurations. This group is intended
to be used by all web-related resources needing access sharing.

The "one domain name, one account" for static sites content works well,
so we adopt it in webserver-alkane.

Units adopted in webserver-alkane will be moved from webserver-legacy
to avoid duplication.

Ref T1828

Test Plan: Deploy to web-001

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1828

Differential Revision: https://devcentral.nasqueron.org/D3034