Align nginx directories
Summary:
As a part of the effort to sync nginx configurations,
both Docker and Alkane platforms will use /var/log/www
to store nginx logs. That will simplify configuration
to send log to OpenSearch and follow the "principle of
less surprise".

The /var/run/web directory is used by php-fpm, so it's best
to move the state in the php-sites unit.

Ref T1828

Test Plan: Deploy on Dwellers and web-001

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1828

Differential Revision: https://devcentral.nasqueron.org/D3037