Don't build PHP 5.6 anymore
Summary:
The support of custom builds allow two things:
  - support a legacy version
  - test a new version still in alpha/dev/RC

The last known *theoretical* uses for an old version of PHP are:
  - Espace Win still contains PHP 3 and PHP 4 code
  - VMailMgr for qmail virtual mailboxes still is PHP 4

There is no actual plan to port VMailMgr to modern PHP,
as we now build mail solutions with Postfix or exim.
And it doesn't run on PHP 5.

Espace Win runs on PHP 5 and early PHP 7 AFAIR, and the engine
can be purely be replaced by modern versions of Keruald/Pluton,
as the code evolved from the Espace Win engine. The other idea
was to create automated tests to help to port, but there is no
concrete plan to do so, and it's not a priority. The porting
itself isn't a priority either.

Also, PHP 5.6 is EOL 4 years ago (2018-12-31).

All other applications run on PHP 7 or PHP 8.

Plan for all the PHP 7 applications was "keep current" with
early reports and patches to help engines and libraries
to support PHP 8.1 and now 8.2. And the PHP 7.4 has been EOL
for now 4 months (2022-12-28).

On a practical basis, the PHP 5 pools are now empty on Ysul
for several years. And PHP 7.x was upgraded to 8.x directly.

Our only real need is so a PHP 8.2 pool.

The custom build mechanism is still needed for PHP 8.3 testing.

Ref T1803

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1803

Differential Revision: https://devcentral.nasqueron.org/D3044