Serve PHP sites on Alkane PaaS
Summary:
Refactor the webserver-legacy php-sites unit, with an effort
to keep in simple.

Enable opcache by default and tune it in production, with larger cache.
This change means php-fpm *MUST* be restarted when a PHP site is updated,
as we don't check the timestamps of the files and so don't invalidate the
cache.

Ref T1828.

Test Plan: Deploy on web-001

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1828

Differential Revision: https://devcentral.nasqueron.org/D3045