Allow updatedb to index ZFS datasets
Summary:
On FreeBSD servers, a weekly task runs updatedb to update the locate index.

Currently, ZFS datasets are excluded from the index, with the expedient
not to index anything with a .zfs subfolder.

An example of the consequence of this default configuration:
`locate index.html` on web-001 doesn't find anything in /var/wwwroot.

Test Plan: Deployed on web-001, run updatedb, files now appears

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3082