Ensure IPv6 routes are announced on OVH
Summary:
OVH network doesn't seem to announce an IPv6 route for a VM at first.

If from the VM, we reach another network, the route is then
announced for a while, before being dropped.

To workaround that behavior, a solution could be to ping regularly
an external site so packets reach OVH router and a route is
announced.

This fix is needed on web-001 for Alkane. Ref T1828.

Test Plan: `ping6 www-alkane.nasqueron.org` now works

Reviewers: dereckson

Reviewed By: dereckson

Tags: #ipv6

Maniphest Tasks: T1828

Differential Revision: https://devcentral.nasqueron.org/D3109