Deploy Orbeon Forms
Summary:
Orbeon Forms allows to fill form, but also to build complex forms.
It's deployed as service for all static and dynamic sites, or for surveys.

It runs on a Tomcat 9 container, and uses a PostgreSQL database on db-A cluster.
SMTP uses an Exim container, in the same network.

**Tomcat users.** This change introduces management for tomcat-users.xml.
This can benefit to future apps needing Tomcat too if we want more monoliths.

**Permissions.** Most parts, form builder, landing pages for form runners, need
an admin access managed at Tomcat level, with credentials stored in Vault.
The form runner allows anonymous access to create a new form.

**Configuration.** Configuration uses direct XML files per instance, located
at roles/paas-docker/containers/files/orbeon/<instance>/ to avoid to maintain
both an YAML pillar to define the config, a XML template with lot of if/for.

Ref T1884

Test Plan: Deploy on Dwellers

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1884

Differential Revision: https://devcentral.nasqueron.org/D3123