Build PostgreSQL with XML support on FreeBSD
Summary:
Orbeon Forms requires XML support. This is an option of the FreeBSD port
not included in the default package.

A bug is open upstream to consider to include XML support in the package,
but meanwhile, we can switch to manual port build.

Fixes T1888.

Test Plan:
  - `make depends` to identify dependencies
  - Build on db-A-001
  - Orbeon Forms can now save forms

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1888

Differential Revision: https://devcentral.nasqueron.org/D3130