Create /var/dataroot for MediaWiki on devserver
Some services like MediaWiki are managed by Salt,
but with a role not available on devserver,
with some directories useful on devserver too.

For example, we don't currently want on WindRiver
a copy...

Summary:
...of all the farm structure, not for my dev
MediaWiki instance, managed a different way
as a standalone service.

Ref T1536

Test Plan: Move mediawiki.dereckso.be to WindRiver

Reviewers: dereckson

Maniphest Tasks: T1536

Differential Revision: https://devcentral.nasqueron.org/D3159