Set privileges for Orbeon forms tables
Summary:
Previously, the Operations Grimoire documented how to manually give rights
to the tables to the relevant user through a SQL query.

While setting rights for Postfix, we discovered in D3252 the syntax
to ask Salt PostgreSQL module to perform that query.

This is a follow-up for a1938c0b3236.

Test Plan: `salt db-A-001 state.apply roles/dbserver-pgsql/content`

Reviewers: DorianWinty

Reviewed By: DorianWinty

Differential Revision: https://devcentral.nasqueron.org/D3253