Provide Python 3 interpreter as python3 executable
Summary:
On FreeBSD, in scenarri where the metaport python3 is missing,
and only python3xx packages are installed, the `python3` symlink can be absent.

As our scripts generally use `#!/usr/bin/env python3` and not a precise shebang
targetting a specific version of Python 3, `python3` should exist and be linked
to the version used by system dependencies.

This fix picks the lower version, as packages tend to be shown in that order.

Test Plan: salt -G 'os:freebsd' state.apply hotfixes/python3

Reviewers: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3288