Ensure webserver-content index can always be regenerated
Summary:
Previously, to regenerate webserver-content index, the index
needed to be existing, so the header can be copied through.

This change allows to regenerate it from scratch.

It also provides a test to ensure the file is up-to-date.

This is a follow-up for 64b8181e37ad.

Test Plan: `rm roles/webserver-content/init.sls && make`

Reviewers: DorianWinty

Differential Revision: https://devcentral.nasqueron.org/D3330