Allow to restart php-fpm on FreeBSD servers after PHP update
Summary:
On FreeBSD, some services can received an extra argument
to launch it several times. We provision a custom php-fpm
service to exactly do that: we want to be able to run several
pools with different PHP versions or configuration.

When the package is reinstalled, for example during a package update,
the custom service is overwritten.

The `restart-php-fpm` command checks the service is the Nasqueron one,
and if not, call Salt to redeploy it. Then, it restart the service.

Test Plan:
Tested on WindRiver with:
  - `pkg update`
  - `pkg upgrade`
  - `restart-php-fpm`

Reviewers: DorianWinty

Reviewed By: DorianWinty

Differential Revision: https://devcentral.nasqueron.org/D3337