Provision /usr/ports a more resiliant way
Summary:
Portsnap is the modern way to provision the FreeBSD ports tree.
The core role takes care of this, but the state was satisfied as
long as the /usr/ports directory exists.

Yet, on Hervil, that directory was present, but empty.
Furthermore, portsnap wasn't install.

This change assures correct handling of that situation
and proper provisioning of the ports tree.

Test Plan: Deployed on Hervil, /usr/ports is now populated with ports.

Reviewers: DorianWinty

Reviewed By: DorianWinty

Differential Revision: https://devcentral.nasqueron.org/D3361