Drop network:ipv6_native from node pillar
Summary:
When migrating the network configuration to NetBox, variables added
to network dictionary for a node needs to be documented there too
or computed by heuristic. As, such, remove unneeded variables is a win.

The ipv6_native variable is:
  - always true when any interface has an IPv6 block
  - always false when no interface ahs an IPv6 block

So, the ipv6_native doesn't bring any other value excepted to avoid
to loop all the interfaces, but for 1-4 interfaces, there is no real
speed gain there.

Ref T1762

Test Plan: Tested on WindRiver

Reviewers: dereckson

Reviewed By: dereckson

Maniphest Tasks: T1762

Differential Revision: https://devcentral.nasqueron.org/D3383