Switch from fixes to flags in node pillar
Summary:
NetBox heuristics to determine if we need to apply IPv6 fix for OVH
is computed at interface level. So, it's easier to check flags rather
than moving flag to a new fixes section.

This change helps to move the node pillar from a static file
to a dynamic NetBox query.

Ref T1762

Test Plan: Deploy on web-001

Reviewers: DorianWinty

Reviewed By: DorianWinty

Maniphest Tasks: T1762

Differential Revision: https://devcentral.nasqueron.org/D3385