Raise Prometheus log level
Summary:
Silence the flood of messages about the storage, focus the log on warnings.

Also fixes the following issue:
Sep 30 19:30:42 windriver prometheus[3333]: Error parsing command line arguments: unrecognized log level "notice"

The level "notice" doesn't exist, it should be "info" in Go
for that level of verbosity. But it's too much information.

This is a follow-up for f2455785.

Test Plan: Deployed on WindRiver

Reviewers: DorianWinty

Reviewed By: DorianWinty

Differential Revision: https://devcentral.nasqueron.org/D3479