Install salt-wrapper on devserver role
Summary:
Once upon a time, the salt-primary role was deployed to Ysul,
which also had the devserver role. The salt-wrapper unit was
naturally add to the Salt part, as it was intended to be used
to select between two primary servers on the same machine.

Since, we're using it for `salt-call --local` too, to quickly
test on devserver the core or devserver role.

Fixes T2042.

Test Plan: Deploy on WindRiver

Reviewers: DorianWinty

Reviewed By: DorianWinty

Maniphest Tasks: T2042

Differential Revision: https://devcentral.nasqueron.org/D3481