Install git everywhere as part of the core role
Summary:
Lot of roles requires Git installation, and it's even used
as an ops tool to create archives of configuration files,
or to pull directly on a server this operations repository
for example.

This change moves Git installation to the core role and
cleans up everywhere else.

This change also provide a configuration for root user,
only if a file doesn't exist to avoid to overwrite [safe] blocks.

Test Plan:
Deployed on WindRiver twice:
  - with /root/.gitconfig existing, only fixed file mode
  - without the file, provisioned it

Reviewers: DorianWinty

Reviewed By: DorianWinty

Differential Revision: https://devcentral.nasqueron.org/D3488