Export metrics for ccache
Summary:
Poudriere speed is improved by ccache use, introduced in D3486.
This change exports metrics for ccache, so we can offer visibility
on the cache current use and efficiency.

From a performance point of view, ccache_cache_hit_rate gives the hit %.

From a monitoring point of view, ccache_cache_size and ccache_max_cache_size
are a nice pair of values to watch, so we know when to increase cache size.

To get metrics for the common cache used by local ports and poudriere:

    $ export CCACHE_DIR=/var/cache/ccache
    $ ccache-metrics

Ref T2039

Test Plan: `ccache-metrics` on WindRiver

Reviewers: DorianWinty

Reviewed By: DorianWinty

Maniphest Tasks: T2039

Differential Revision: https://devcentral.nasqueron.org/D3492