Create datacube for FreeBSD Nasqueron packages repository
Summary:
Role freebsd-repo uses the /var/repo directory, a prime candidate for a datacube
ZFS volume in the devserver context.

The volume belongs to the "builder" user per freebsd-repo role.

The parent volume "packages" allows to create a hierarchy for the different
directories in packages.nasqueron.org. This change takes care of the FreeBSD one.

Ref T1850

Test Plan: Deployed on WindRiver

Reviewers: DorianWinty

Reviewed By: DorianWinty

Maniphest Tasks: T1850

Differential Revision: https://devcentral.nasqueron.org/D3495