Allow all users to read RabbitMQ credential to get notifications
Summary:
The RabbitMQ credential for notifications is intended to be shared
among devserver users, as such, it makes sense to make the secret
file readable for the nasquenautes group.

This credential allows only to READ from a queue on the dev vhost,
and doesn't give access to anything else than notifications.

Test Plan: Deployed on WindRiver

Reviewers: aceppaluni, dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3524