Write log by nginx subdomain
Summary:
Some of the subdomains didn't define logging directive in server blocks.
That led to logs reaching a general unified log, not the more convenient
to tail errors for a specific site when working on new development or trying
to debug issues.

Test Plan: Deploy to web-001, windriver and hervil

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3555