Select 172.27.27. address when resolving private network
Summary:
On Windriver, the first interface ix0 is configured for RPN connectivity
with a private address. That address is so used in node.resolve_network.

This change requires private network configuration to be an interface
where the first IPv4 address starts by 172.27.27. The goal is to allow services
to listen to that IP address.

CIDR computation is not currently required, as we use a /24. The code comment
will warn to change the code if we switch to something else than /8 /16 /24.

Reviewers: DorianWinty

Reviewed By: DorianWinty

Differential Revision: https://devcentral.nasqueron.org/D3594