Upgrade zenerre to PHP 8
Summary:
Auth Grove still uses PHP 7. This is the reason this agent was still 7.4.

If we decide to use Auth Grove (see T1771), the first priority
will be to move our classes to a fresh PHP 8.4 / Laravel 11 installation.

There is also a push to use clean architecture and separate the framework layer
from the code layer.

As all other applications use PHP 8, it makes no sense to pin that version.

Test Plan: Deploy to docker-002

Reviewers: dereckson

Reviewed By: dereckson

Subscribers: DorianWinty, aceppaluni

Differential Revision: https://devcentral.nasqueron.org/D3606