Allow to discover IPv6 gateway for Online servers
Summary:
Online doesn't document gateway for IPv6 blocks.

Normally, `inet6 -ifdisabled accept_rtadv up` and the DHCPv6 request
to authenticate the right to use the block with DUID should be enough.

If we want to explicitly set a gateway in the routing table, we can
discover it through neighbour discovery protocol (NDP): this will be
the only IP adress not associated to a known MAC address.

With that informatino, gateway can then be declared on NetBox
by an ops running the query-ndp-for-ipv6-neighbor.sh script.

Test Plan: `sh utils/query-ndp-for-ipv6-neighbor.sh` on WindRiver

Reviewers: DorianWinty

Reviewed By: DorianWinty

Differential Revision: https://devcentral.nasqueron.org/D3651