Fix node.resolve_network tests
Summary:
Any private IP addressi instead of Drake network is ignored
by the node.resolve_network function, so we need to update
the private address in test data to belong to that range.

Furthermore, new entries have been added to the dictionary,
not handled by current test suite, so it makes more sense
to test for an exact subset than for equality.

This is a follow-up for 5877d5111f9a.
This is a follow-up for 4d08af81970d.

Ref T2123

Test Plan: `(cd _tests && python3 -m unittest modules/test_node.py)`

Reviewers: DorianWinty

Reviewed By: DorianWinty

Maniphest Tasks: T2123

Differential Revision: https://devcentral.nasqueron.org/D3663