Allow Phabricator to store files on local disk
Summary:
Currently, Phabricator stores files in database. This become to be tricky
for efficient MySQL databases backups. Plan is to move to local storage.

This change prepares a migration path by creating a directory to store
the files and declaring it in the configuration.

References:
  - https://secure.phabricator.com/book/phabricator/article/configuring_file_storage/

Test Plan:
Applied on docker-002

Files migrated for DevCentral instance

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3730