Add interwiki agora: and wolfplex:
Summary:
Create a more restricted user for MediaWiki maintenance.
Privileges are limited to the affected tables.

Populate a support SQL script with the interwiki list.
The use of REPLACE allows to avoid to read tables first;
the script is safe to run each time it's modified.

Install MariaDB client on saas-mediawiki role,
so that script can be run from there.

References:
  - https://www.mediawiki.org/wiki/Manual:Interwiki
  - https://www.mediawiki.org/wiki/Manual:Interwiki_table

Test Plan:
  - Create a link `[[agora:Operations grimoire]]` on Wolfplex wiki

Reviewers: dereckson

Reviewed By: dereckson

Differential Revision: https://devcentral.nasqueron.org/D3736